+++ scripts/lib/utilities/activityUtils.js
@@ -134,6 +134,15 @@ function canUse(activity) {
     }
     return true;
 }
+async function correctSpellLink(activity, spell) {
+    return await genericUtils.update(activity, {'spell.uuid': spell.uuid});
+}
+function isHidden(activity) {
+    if (activity.midiProperties.automationOnly) return true;
+    let cprRiders = genericUtils.getProperty(activity.item, 'flags.chris-premades.hiddenActivities') ?? [];
+    cprRiders = cprRiders.map(i => activityUtils.getActivityByIdentifier(activity.item, i)?.id).filter(i => i);
+    return cprRiders.includes(activity.id);
+}
 export let activityUtils = {
     getActivityByIdentifier,
     getIdentifier,
@@ -144,5 +153,7 @@ export let activityUtils = {
     getMod,
     hasSave,
     isSpellActivity,
-    canUse
+    canUse,
+    correctSpellLink,
+    isHidden
 };